refactor: member 및 party 도메인의 lookup 서비스를 타 도메인과 동일하게 query 아래로 이동
