+++ src/main/java/umc/cockple/demo/domain/exercise/service/command/ExerciseLifecycleCommandService.java
@@ -17,10 +17,10 @@ import umc.cockple.demo.domain.exercise.service.command.model.ExerciseUpdateAddr
 import umc.cockple.demo.domain.exercise.service.command.model.ExerciseUpdateCommand;
 import umc.cockple.demo.domain.exercise.service.support.reader.ExerciseReader;
 import umc.cockple.demo.domain.member.domain.Member;
-import umc.cockple.demo.domain.member.service.support.MemberLookupService;
+import umc.cockple.demo.domain.member.service.query.lookup.MemberLookupService;
 import umc.cockple.demo.domain.party.domain.Party;
 import umc.cockple.demo.domain.party.repository.PartyRepository;
-import umc.cockple.demo.domain.party.service.support.PartyLookupService;
+import umc.cockple.demo.domain.party.service.query.lookup.PartyLookupService;
 
 @Service
 @Transactional
